+++ src/extension.ts
@@ -9,22 +9,26 @@ import { MediaFileType } from './interfaces'
 import { printToChannel} from './utils'
 import Converter from './Converter'
 import ConverterImg from './ConverterImg'
+import TreeViewProvider from './batchTreeview'
 
 ffmpeg.setFfmpegPath(pathToFfmpeg!)
 const { showErrorMessage, showInformationMessage } = window
 const { MP3, MP4, JPG, WAV } = MediaFileType
 
 export function activate(context: ExtensionContext) {
-  const rc = commands.registerCommand
   init()
-
+  const treeViewProvider = new TreeViewProvider()
+  window.registerTreeDataProvider('emcTreeView', treeViewProvider)
+  const rc = commands.registerCommand
   context.subscriptions.concat([
     rc('emc.convertMp3', (uri: Uri) => Converter.convert(uri, MP3)),
     rc('emc.convertMp4', (uri: Uri) => Converter.convert(uri, MP4)),
     rc('emc.convertJpg', (uri: Uri) => ConverterImg.convert(uri, JPG)),
     rc('emc.convertWav', (uri: Uri) => Converter.convert(uri, WAV)),
     rc('emc.download', download),
-    rc('emc.revealFfmpegBin', revealFfmpegBin)
+    rc('emc.revealFfmpegBin', revealFfmpegBin),
+    rc('emc.addToQueue', (file: string) => treeViewProvider.addToQueue(file)),
+    rc('emc.removeFromQueue', (file: string) => treeViewProvider.removeFromQueue(file))
   ])
 }